+++ backend/src/baserow/contrib/database/fields/fields.py
@@ -1,0 +1,21 @@
+from django.db import models
+from django.db.models.fields.related_descriptors import ForwardManyToOneDescriptor
+
+
+class SingleSelectForwardManyToOneDescriptor(ForwardManyToOneDescriptor):
+    def get_object(self, instance):
+        """
+        Tries to fetch the reference object, but if it fails because it doesn't exist,
+        the value will be set to None instead of failing hard.
+        """
+
+        try:
+            return super().get_object(instance)
+        except self.field.remote_field.model.DoesNotExist:
+            setattr(instance, self.field.name, None)
+            instance.save()
+            return None
+
+
+class SingleSelectForeignKey(models.ForeignKey):
+    forward_related_accessor_class = SingleSelectForwardManyToOneDescriptor