+++ backend/src/baserow/contrib/database/views/handler.py
@@ -459,20 +459,28 @@ class ViewHandler:
 
         order_by = []
 
-        for view_filter in view.viewsort_set.all():
+        for view_sort in view.viewsort_set.all():
             # If the to be sort field is not present in the `_field_objects` we
             # cannot filter so we raise a ValueError.
-            if view_filter.field_id not in model._field_objects:
+            if view_sort.field_id not in model._field_objects:
                 raise ValueError(f'The table model does not contain field '
-                                 f'{view_filter.field_id}.')
+                                 f'{view_sort.field_id}.')
 
-            field_name = model._field_objects[view_filter.field_id]['name']
-            order = F(field_name)
+            field = model._field_objects[view_sort.field_id]['field']
+            field_name = model._field_objects[view_sort.field_id]['name']
+            field_type = model._field_objects[view_sort.field_id]['type']
 
-            if view_filter.order == 'ASC':
-                order = order.asc(nulls_first=True)
-            else:
-                order = order.desc(nulls_last=True)
+            order = field_type.get_order(field, field_name, view_sort)
+
+            # If the field type does not have a specific ordering expression we can
+            # order the default way.
+            if not order:
+                order = F(field_name)
+
+                if view_sort.order == 'ASC':
+                    order = order.asc(nulls_first=True)
+                else:
+                    order = order.desc(nulls_last=True)
 
             order_by.append(order)