+++ backend/src/baserow/contrib/database/views/view_filters.py
@@ -6,12 +6,13 @@ from dateutil import parser
 from dateutil.parser import ParserError
 
 from django.db.models import Q, IntegerField, BooleanField
-from django.db.models.fields.related import ManyToManyField
+from django.db.models.fields.related import ManyToManyField, ForeignKey
 from django.contrib.postgres.fields import JSONField
 
 from baserow.contrib.database.fields.field_types import (
     TextFieldType, LongTextFieldType, URLFieldType, NumberFieldType, DateFieldType,
-    LinkRowFieldType, BooleanFieldType, EmailFieldType, FileFieldType
+    LinkRowFieldType, BooleanFieldType, EmailFieldType, FileFieldType,
+    SingleSelectFieldType
 )
 
 from .registries import ViewFilterType
@@ -204,6 +205,33 @@ class DateNotEqualViewFilterType(NotViewFilterTypeMixin, DateEqualViewFilterType
     type = 'date_not_equal'
 
 
+class SingleSelectEqualViewFilterType(ViewFilterType):
+    """
+    The single select equal filter accepts a select option id as filter value. This
+    filter is only compatible with the SingleSelectFieldType field type.
+    """
+
+    type = 'single_select_equal'
+    compatible_field_types = [SingleSelectFieldType.type]
+
+    def get_filter(self, field_name, value, model_field):
+        value = value.strip()
+
+        if value == '':
+            return Q()
+
+        try:
+            int(value)
+            return Q(**{f'{field_name}_id': value})
+        except Exception:
+            return Q()
+
+
+class SingleSelectNotEqualViewFilterType(NotViewFilterTypeMixin,
+                                         SingleSelectEqualViewFilterType):
+    type = 'single_select_not_equal'
+
+
 class BooleanViewFilterType(ViewFilterType):
     """
     The boolean filter tries to convert the provided filter value to a boolean and
@@ -253,12 +281,16 @@ class EmptyViewFilterType(ViewFilterType):
         DateFieldType.type,
         LinkRowFieldType.type,
         EmailFieldType.type,
-        FileFieldType.type
+        FileFieldType.type,
+        SingleSelectFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field):
         # If the model_field is a ManyToMany field we only have to check if it is None.
-        if isinstance(model_field, ManyToManyField):
+        if (
+            isinstance(model_field, ManyToManyField) or
+            isinstance(model_field, ForeignKey)
+        ):
             return Q(**{f'{field_name}': None})
 
         if isinstance(model_field, BooleanField):