+++ web-frontend/modules/core/pages/styleGuide.vue
@@ -234,6 +234,43 @@
             </div>
           </div>
           <div class="control">
+            <label class="control__label">
+              Dropdown by link {{ dropdown }}
+            </label>
+            <div class="control__elements">
+              <div style="width: 200px;">
+                <a
+                  ref="dropdownLink"
+                  @click="$refs.dropdown1.toggle($refs.dropdownLink)"
+                  >Open dropdown</a
+                >
+                <Dropdown
+                  ref="dropdown1"
+                  v-model="dropdown"
+                  :show-input="false"
+                >
+                  <DropdownItem name="Choice 1" value="choice-1"></DropdownItem>
+                  <DropdownItem
+                    name="Choice 2"
+                    value="choice-2"
+                    icon="pencil"
+                  ></DropdownItem>
+                  <DropdownItem
+                    name="Choice 3"
+                    value="choice-3"
+                    icon="database"
+                  ></DropdownItem>
+                  <DropdownItem
+                    name="Choice 4"
+                    value="choice-4"
+                    icon="times"
+                    :disabled="true"
+                  ></DropdownItem>
+                </Dropdown>
+              </div>
+            </div>
+          </div>
+          <div class="control">
             <label class="control__label">Datepicker field</label>
             <div class="control__elements">
               value: {{ date }}