+++ web-frontend/modules/database/components/field/FieldSingleSelectDropdownItem.vue
@@ -1,6 +1,6 @@
 <template>
   <li
-    class="select__item"
+    class="field-single-select__dropdown-item"
     :class="{
       hidden: !isVisible(query),
       active: isActive(value),
@@ -9,16 +9,16 @@
     }"
   >
     <a
-      class="select__item-link"
+      class="field-single-select__dropdown-link"
       @click="select(value, disabled)"
       @mousemove="hover(value, disabled)"
     >
-      <i
-        v-if="icon"
-        class="select__item-icon fas fa-fw"
-        :class="'fa-' + icon"
-      ></i>
-      {{ name }}
+      <div
+        class="field-single-select__dropdown-option"
+        :class="'background-color--' + color"
+      >
+        {{ name }}
+      </div>
     </a>
   </li>
 </template>
@@ -27,7 +27,14 @@
 import dropdownItem from '@baserow/modules/core/mixins/dropdownItem'
 
 export default {
-  name: 'DropdownItem',
+  name: 'FieldSingleSelectDropdownItem',
   mixins: [dropdownItem],
+  props: {
+    color: {
+      type: String,
+      required: false,
+      default: '',
+    },
+  },
 }
 </script>